third_party/blink/web_tests/external/wpt/webrtc-encoded-transform/script-transform.https-expected.txt

This is a testharness.js-based test.
[FAIL] transform messaging
  promise_test: Unhandled rejection with value: object "ReferenceError: RTCRtpScriptTransform is not defined"
[FAIL] Cannot reuse attached transforms
  promise_test: Unhandled rejection with value: object "ReferenceError: RTCRtpScriptTransform is not defined"
[FAIL] audio exchange with transform
  promise_test: Unhandled rejection with value: object "ReferenceError: RTCRtpScriptTransform is not defined"
[FAIL] video exchange with transform
  promise_test: Unhandled rejection with value: object "ReferenceError: RTCRtpScriptTransform is not defined"
Harness: the test ran to completion.
